requirements.txt
black==21.12b0
selenium==4.1.0
python-dotenv==0.19.2
